Adjusted repo to split between model and dtos
