+++ src/resources/movie/repos/IMovieRepository.ts
@@ -1,9 +1,10 @@
+import { GENRES } from '../../../shared/enums/genres';
 import { IGetMoviesDto } from '../dtos/getMovies.dto';
-import { IMovieDto } from '../dtos/movie.dto';
-import { IMovie, IMovieWithId } from '../movie.model';
+import { IMovieDto, IMovieDtoWithId } from '../dtos/IMovieDto';
+import { IMovieWithId } from '../movie.model';
 
 export default interface IMovieRepository {
-    create(movie: IMovieWithId): Promise<IMovieDto>;
+    create(movie: IMovieWithId): Promise<IMovieDtoWithId>;
     getMoviesByDuration(from: number, to: number): Promise<IGetMoviesDto>;
     getMoviesByGenres(genres: GENRES[]): Promise<IGetMoviesDto>;
     getMoviesByGenresAndDuration(genres: GENRES[], duration: {from: number, to: number}): Promise<IGetMoviesDto>;
